requirements.txt
pymongo
telepot
pyyaml
